move: CartPage용 컴포넌트를 별도 파일에 선언하기
