+++ src/refactoring/components/CartPage/ProductList.tsx
@@ -1,5 +1,5 @@
 import ProductListItem from './ProductListItem';
-import type { Product } from '../../types';
+import type { Product } from '../../../types';
 
 interface ProductListProps {
   products: Product[];
